Changed calculator to work on ints

+++ include/calculator/Calculator.h
@@ -2,8 +2,8 @@
 
 class Calculator {
 public:
-    double add(double a, double b);
-    double subtract(double a, double b);
-    double multiply(double a, double b);
-    double divide(double a, double b);
+    int add(int a, int b);
+    int subtract(int a, int b);
+    int multiply(int a, int b);
+    int divide(int a, int b);
 };
